Calculator divide수정

+++ src/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
         return a*b;
     }
 
-    int divide(int a, int b){
-        return a/b;
+    float divide(int a, int b){
+        return (float)a/b;
     }
 }
